Update README.md
Your team will develop a software for RU Pizza to manage the pizza orders. Your team must use JavaFX to develop the GUIs for taking orders, placing orders, and cancelling orders. The pizzeria offers five specialty pizzas, Deluxe, Supreme, Meatzza, Seafood and Pepperoni, with store customized toppings as shown in the table below. The pizzeria also offers “build your own” pizzas, which allow the customers to choose the toppings they like, up to 7 toppings. The store charges $1 each for extra cheese and extra sauce. The store staff will be the one using the software to take the orders from the customers. Each order is uniquely identified by an order number. For simplicity, the software system will not keep track of the dates and customer information of the orders.

When build your own pizza is selected, the system shall provide 13 toppings for customization.

When the user is customizing a build your own pizza, the user can add/remove toppings, with a maximum of 7 toppings, and the user can select tomato sauce or alfredo sauce.

When the user is ordering specialty pizzas, the system shall provide the five specialty pizza options, Deluxe, Supreme, Meatzza, Pepperoni and Seafood.

When a specialty pizza is selected, the system shall display the image of the specialty pizza and the store customized toppings. The user cannot change the toppings, but can select, size, extra sauce and extra cheese.